refactor: move all data types into a subfolder 'base/'
Having everything is a single folder makes it very difficult to differentiate
between context and use-case.
