+++ rt/src/base/Atom.mts
@@ -1,7 +1,7 @@
-import runId from "./runId.mjs"
+import runId from "../runId.mjs"
 import { TroupeType } from "./TroupeTypes.mjs"
 import { TroupeRawValue } from "./TroupeRawValue.mjs";
-import * as levels from './Level.mjs'
+import * as levels from '../Level.mjs'
 
 let rt_uuid = runId
