+++ rt/src/base/Authority.mts
@@ -1,4 +1,4 @@
-import { Level, BOT } from "./Level.mjs";
+import { Level, BOT } from "../Level.mjs";
 
 import { TroupeRawValue } from "./TroupeRawValue.mjs";
 import { TroupeType } from "./TroupeTypes.mjs";
