+++ rt/src/base/LVal.mts
@@ -1,5 +1,5 @@
-import * as levels from './Level.mjs'
-import { Level } from "./Level.mjs";
+import * as levels from '../Level.mjs'
+import { Level } from "../Level.mjs";
 import * as Ty from './TroupeTypes.mjs';
 import { TroupeRawValue } from './TroupeRawValue.mjs';
