+++ rt/src/base/RawClosure.mts
@@ -1,6 +1,6 @@
 import {TroupeAggregateRawValue} from './TroupeRawValue.mjs'
 import {ClosureType, TroupeType} from './TroupeTypes.mjs'
-import * as levels from './Level.mjs' 
+import * as levels from '../Level.mjs' 
 export function RawClosure (e, _t, f) : TroupeAggregateRawValue { 
   let closure:any = () => {    
     return f (e)
