+++ rt/src/base/RawList.mts
@@ -1,8 +1,8 @@
 import {TroupeType} from './TroupeTypes.mjs'
 import {TroupeAggregateRawValue} from './TroupeRawValue.mjs'
 import {LVal, listStringRep} from './LVal.mjs'
-import { Level } from './Level.mjs'
-import * as levels from './Level.mjs'
+import { Level } from '../Level.mjs'
+import * as levels from '../Level.mjs'
 
 
 export abstract class RawList implements TroupeAggregateRawValue {
