+++ rt/src/base/RawRecord.mts
@@ -1,9 +1,9 @@
 import {TroupeType} from './TroupeTypes.mjs'
 import {TroupeAggregateRawValue} from './TroupeRawValue.mjs'
 import {LVal, listStringRep} from './LVal.mjs'
-import { Level } from './Level.mjs'
-import * as levels from './Level.mjs'
-import { assertIsRecord } from './Asserts.mjs'
+import { Level } from '../Level.mjs'
+import * as levels from '../Level.mjs'
+import { assertIsRecord } from '../Asserts.mjs'
 
 
 export class RawRecord implements TroupeAggregateRawValue {
