+++ rt/src/base/UnitVal.mts
@@ -1,5 +1,5 @@
 import RawUnit from './RawUnit.mjs'
-import { BOT } from './Level.mjs';
+import { BOT } from '../Level.mjs';
 import { LVal } from './LVal.mjs';
 
 export let __unit = new LVal (RawUnit, BOT, BOT)
