+++ rt/src/builtins/UserRuntimeZero.mts
@@ -1,21 +1,21 @@
 'use strict'
-import { runtimeEquals } from '../EqualityChecker.mjs'
-import { isListFlagSet, isTupleFlagSet, mkTuple, mkList, mkWithRecord, mkRecord, isRecord } from '../ValuesUtil.mjs'
-import { LVal, LValCopyAt, LCopyVal } from '../LVal.mjs'
-import { Nil, Cons, RawList } from '../RawList.mjs'
+import { runtimeEquals } from '../base/EqualityChecker.mjs'
+import { isListFlagSet, isTupleFlagSet, mkTuple, mkList, mkWithRecord, mkRecord, isRecord } from '../base/ValuesUtil.mjs'
+import { LVal, LValCopyAt, LCopyVal } from '../base/LVal.mjs'
+import { Nil, Cons, RawList } from '../base/RawList.mjs'
 import { loadLibsAsync } from '../loadLibsAsync.mjs';
 import * as levels from '../Level.mjs'
-import { BaseFunctionWithExplicitArg, ServiceFunction } from '../BaseFunction.mjs'
-import { Atom } from '../Atom.mjs'
-import { __unit } from '../UnitVal.mjs'
+import { BaseFunctionWithExplicitArg, ServiceFunction } from '../base/BaseFunction.mjs'
+import { Atom } from '../base/Atom.mjs'
+import { __unit } from '../base/UnitVal.mjs'
 import { RuntimeInterface } from '../RuntimeInterface.mjs';
-import { RawRecord } from '../RawRecord.mjs'
-import { TroupeType } from '../TroupeTypes.mjs'
-import { RawClosure } from '../RawClosure.mjs'
-import RawUnit from '../RawUnit.mjs'
+import { RawRecord } from '../base/RawRecord.mjs'
+import { TroupeType } from '../base/TroupeTypes.mjs'
+import { RawClosure } from '../base/RawClosure.mjs'
+import RawUnit from '../base/RawUnit.mjs'
 import { Thread } from '../Thread.mjs'
-import { TroupeRawValue } from '../TroupeRawValue.mjs'
-import { RawTuple } from '../RawTuple.mjs'
+import { TroupeRawValue } from '../base/TroupeRawValue.mjs'
+import { RawTuple } from '../base/RawTuple.mjs'
 import { Level } from '../Level.mjs'
 import { rawAssertNotZero } from '../Asserts.mjs'
