+++ rt/src/builtins/math.mts
@@ -1,5 +1,5 @@
 import {UserRuntimeZero, Constructor, mkBase} from './UserRuntimeZero.mjs'
-import { LVal } from '../LVal.mjs';
+import { LVal } from '../base/LVal.mjs';
 import * as levels  from '../Level.mjs'
 import { assertIsUnit, assertIsNumber } from '../Asserts.mjs'
