+++ rt/src/builtins/monitor.mts
@@ -1,6 +1,6 @@
 import {UserRuntimeZero, Constructor, mkBase} from './UserRuntimeZero.mjs'
 import { assertNormalState, assertIsProcessId, assertIsString } from '../Asserts.mjs'
-import { __unit } from '../UnitVal.mjs';
+import { __unit } from '../base/UnitVal.mjs';
 
 export function BuiltinMonitors <TBase extends Constructor<UserRuntimeZero>> (Base:TBase) {
