+++ rt/src/builtins/sandbox.mts
@@ -1,13 +1,13 @@
 import { UserRuntimeZero, Constructor, mkBase } from './UserRuntimeZero.mjs'
-import { LVal, LValCopyAt } from '../LVal.mjs';
+import { LVal, LValCopyAt } from '../base/LVal.mjs';
 import { assertNormalState, assertIsNTuple, assertIsNumber, assertIsFunction } from '../Asserts.mjs'
-import { __unit } from '../UnitVal.mjs';
+import { __unit } from '../base/UnitVal.mjs';
 import { getCliArgs, TroupeCliArg } from '../TroupeCliArgs.mjs';
 import { mkLogger } from '../logger.mjs'
 import SandboxStatus from '../SandboxStatus.mjs'
-import { mkTuple } from '../ValuesUtil.mjs';
+import { mkTuple } from '../base/ValuesUtil.mjs';
 import { CALLSIZE, RETOFFSET, Thread } from '../Thread.mjs';
-import { RawClosure, SandboxResumption } from '../RawClosure.mjs';
+import { RawClosure, SandboxResumption } from '../base/RawClosure.mjs';
 import { RuntimeInterface } from '../RuntimeInterface.mjs';
 
 const argv = getCliArgs();
