+++ rt/src/builtins/service.mts
@@ -1,5 +1,5 @@
 import {UserRuntimeZero, Constructor, mkBase, mkService} from './UserRuntimeZero.mjs'
-import { __unit } from '../UnitVal.mjs'
+import { __unit } from '../base/UnitVal.mjs'
 export function BuiltinService <TBase extends Constructor<UserRuntimeZero>> (Base:TBase) {
     return class extends Base {
         _servicetest = mkService(() => this.runtime.$service.servicetest(),"servicetest")
