+++ rt/src/builtins/whereis.mts
@@ -1,14 +1,14 @@
 'use strict'
 import { UserRuntimeZero, Constructor, mkBase } from './UserRuntimeZero.mjs'
 import * as levels from '../Level.mjs'
-import { ProcessID } from '../process.mjs';
+import { ProcessID } from '../base/process.mjs';
 const { lub, flowsTo } = levels
 import {deserialize} from '../deserialize.mjs'
 import { __nodeManager } from '../NodeManager.mjs';
 import { assertNormalState, assertIsNTuple, assertIsString, assertIsProcessId, assertIsAuthority, assertIsRootAuthority, assertIsNode } from '../Asserts.mjs';
-import { __unit } from '../UnitVal.mjs';
+import { __unit } from '../base/UnitVal.mjs';
 import { nodeTrustLevel } from '../TrustManager.mjs';
-import { LVal } from '../LVal.mjs'
+import { LVal } from '../base/LVal.mjs'
 import {p2p} from '../p2p/p2p.mjs'
 import { getCliArgs, TroupeCliArg } from '../TroupeCliArgs.mjs';
 import { mkLogger } from '../logger.mjs'
